Refactor server structure and configuration:
- Remove `main.py` and split functionality into `app.py` and `build.py`
- Simplify `ServerBaseSettings` in `config.py`:
  - Remove `APP_TYPE`, `DEFAULT_PORT`, and optional `PORT`
  - Set default `PORT` to 8680
- Delete `LlmSettings` and `AgentSettings` classes
- Update `create_app` and `run_app` functions in `build.py`
- Add health check endpoint
- Implement `app_lifespan` context manager
- Remove model registration and discovery logic
